a few examples
