+++ README.md
@@ -1,0 +1,15 @@
+
+A simple web application to help other learn Flask.
+
+Run on Linux as follows:
+
+```
+export FLASK_APP=main.py
+flask run
+```
+
+Run on Windows as follows:
+```
+set FLASK_APP=main.py
+flask run
+```
